+++ compiler-core/src/parse/tests.rs
@@ -354,6 +354,215 @@ fn triple_equals_with_whitespace() {
     );
 }
 
+#[test]
+fn merge_conflict_equals() {
+    assert_error!(
+        "let wobble: Int = 32
+        =======",
+        ParseError {
+            error: ParseErrorType::LexError {
+                error: LexicalError {
+                    error: LexicalErrorType::MergeConflictIndicator,
+                    location: SrcSpan { start: 29, end: 36 },
+                }
+            },
+            location: SrcSpan { start: 29, end: 36 },
+        }
+    );
+}
+
+#[test]
+fn merge_conflict_equals_snapshot() {
+    assert_error!(
+        "let wobble: Int = 32
+        ======="
+    );
+}
+
+#[test]
+fn six_equals_not_merge_conflict() {
+    assert_error!(
+        "let wobble: Int = 32
+        ======",
+        ParseError {
+            error: ParseErrorType::OpNakedRight,
+            location: SrcSpan { start: 29, end: 31 },
+        }
+    );
+}
+
+#[test]
+fn merge_conflict_lt() {
+    assert_error!(
+        "let wobble: Int = 32
+        <<<<<<<",
+        ParseError {
+            error: ParseErrorType::LexError {
+                error: LexicalError {
+                    error: LexicalErrorType::MergeConflictIndicator,
+                    location: SrcSpan { start: 29, end: 36 },
+                }
+            },
+            location: SrcSpan { start: 29, end: 36 },
+        }
+    );
+}
+
+#[test]
+fn merge_conflict_lt_with_head() {
+    assert_error!(
+        "let wobble: Int = 32
+        <<<<<<< HEAD",
+        ParseError {
+            error: ParseErrorType::LexError {
+                error: LexicalError {
+                    error: LexicalErrorType::MergeConflictIndicator,
+                    location: SrcSpan { start: 29, end: 36 },
+                }
+            },
+            location: SrcSpan { start: 29, end: 36 },
+        }
+    );
+}
+
+#[test]
+fn merge_conflict_lt_with_head_snapshot() {
+    assert_error!(
+        "let wobble: Int = 32
+        <<<<<<< HEAD"
+    );
+}
+
+#[test]
+fn four_lt_not_merge_conflict() {
+    assert_error!(
+        "let wobble: Int = 32
+        <<<<",
+        ParseError {
+            error: ParseErrorType::UnexpectedEof,
+            location: SrcSpan { start: 0, end: 0 },
+        }
+    );
+}
+
+#[test]
+fn five_lt_not_merge_conflict() {
+    assert_error!(
+        "let wobble: Int = 32
+        <<<<<",
+        ParseError {
+            error: ParseErrorType::UnexpectedToken {
+                token: Token::Less,
+                expected: vec!["`>>`".into(), "a bit array segment".into()],
+                hint: None,
+            },
+            location: SrcSpan { start: 34, end: 35 },
+        }
+    );
+}
+
+#[test]
+fn six_lt_not_merge_conflict() {
+    assert_error!(
+        "let wobble: Int = 32
+        <<<<<<",
+        ParseError {
+            error: ParseErrorType::UnexpectedEof,
+            location: SrcSpan { start: 0, end: 0 },
+        }
+    );
+}
+
+#[test]
+fn merge_conflict_gt() {
+    assert_error!(
+        "let wobble: Int = 32
+        >>>>>>>",
+        ParseError {
+            error: ParseErrorType::LexError {
+                error: LexicalError {
+                    error: LexicalErrorType::MergeConflictIndicator,
+                    location: SrcSpan { start: 29, end: 36 },
+                }
+            },
+            location: SrcSpan { start: 29, end: 36 },
+        }
+    );
+}
+
+#[test]
+fn merge_conflict_gt_with_branch_name() {
+    assert_error!(
+        "let wobble: Int = 32
+        >>>>>>> main",
+        ParseError {
+            error: ParseErrorType::LexError {
+                error: LexicalError {
+                    error: LexicalErrorType::MergeConflictIndicator,
+                    location: SrcSpan { start: 29, end: 36 },
+                }
+            },
+            location: SrcSpan { start: 29, end: 36 },
+        }
+    );
+}
+
+#[test]
+fn merge_conflict_gt_with_branch_name_snapshot() {
+    assert_error!(
+        "let wobble: Int = 32
+        >>>>>>> main"
+    );
+}
+
+#[test]
+fn four_gt_not_merge_conflict() {
+    assert_error!(
+        "let wobble: Int = 32
+        >>>>",
+        ParseError {
+            error: ParseErrorType::UnexpectedToken {
+                token: Token::GtGt,
+                expected: vec!["An import, const, type, or function.".into()],
+                hint: None,
+            },
+            location: SrcSpan { start: 29, end: 31 },
+        }
+    );
+}
+
+#[test]
+fn five_gt_not_merge_conflict() {
+    assert_error!(
+        "let wobble: Int = 32
+        >>>>>",
+        ParseError {
+            error: ParseErrorType::UnexpectedToken {
+                token: Token::GtGt,
+                expected: vec!["An import, const, type, or function.".into()],
+                hint: None,
+            },
+            location: SrcSpan { start: 29, end: 31 },
+        }
+    );
+}
+
+#[test]
+fn six_gt_not_merge_conflict() {
+    assert_error!(
+        "let wobble: Int = 32
+        >>>>>>",
+        ParseError {
+            error: ParseErrorType::UnexpectedToken {
+                token: Token::GtGt,
+                expected: vec!["An import, const, type, or function.".into()],
+                hint: None
+            },
+            location: SrcSpan { start: 29, end: 31 },
+        }
+    );
+}
+
 // https://github.com/gleam-lang/gleam/issues/1231
 #[test]
 fn pointless_spread() {