+++ CHANGELOG.md
@@ -107,6 +107,10 @@
   when building large projects.
   ([John Downey](https://github.com/jtdowney))
 
+- The compiler now provides a more specific error message when encountering a
+  merge conflict indicator
+  ([0xda157](https://github.com/0xda157))
+
 ### Language server
 
 - The language server now supports go-to-definition, find-references and rename